Fixed compiler warning

+++ dash/test/MatrixTest.cc
@@ -14,8 +14,8 @@ TEST_F(MatrixTest, OddSize)
 
   dash::Matrix<int, 2, index_t, pattern_t> matrix(dash::SizeSpec<2>(8, 15));
 
-  for (size_t i = 0; i < matrix.extent(0); i++) {
-    for (size_t j = 0; j < matrix.extent(1); j++) {
+  for (int i = 0; i < matrix.extent(0); i++) {
+    for (int j = 0; j < matrix.extent(1); j++) {
       if (matrix(i,j).is_local()) {
         DASH_LOG_TRACE("MatrixText.OddSize", "(", i, ",", j, ")",
                        "unit:", _dash_id);
